+++ src/Contract/Parser.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Duon\Sire\Contract;
+
+/** @api */
+interface Parser
+{
+	/**
+	 * @return array<array-key, mixed>
+	 * @throws \Duon\Sire\Exception\ValidationError
+	 */
+	public function parse(array $data): array;
+}